# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Playbook {
  id: number;
  name: string;
  description: string;
  qualityScore?: number | null;
  emailCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface Pattern {
  id: number;
  playbookId: number;
  type: string;
  text: string;
  examples: string[];
}

export interface PlaybookWithPatterns {
  id: number;
  name: string;
  description: string;
  qualityScore?: number | null;
  emailCount: number;
  createdAt: string;
  updatedAt: string;
  patterns: Pattern[];
  principles: string[];
}

export interface CreatePlaybookBody {
  name: string;
  description: string;
  icpVerticals?: string[];
  icpPersonas?: string[];
  icpPainPoints?: string[];
  icpDifferentiators?: string[];
  icpProofPoints?: string[];
  icpCompanySize?: string | null;
}

export interface AnalyzeEmailsBody {
  emails: string[];
}

export type AnalysisResultPatternsItem = {
  type: string;
  text: string;
  examples: string[];
};

export interface AnalysisResult {
  patterns: AnalysisResultPatternsItem[];
  principles: string[];
  qualityScore: number;
}

export interface Generation {
  id: number;
  type: string;
  company?: string | null;
  role?: string | null;
  output: string;
  playbookId?: number | null;
  createdAt: string;
}

export interface GenerationResult {
  output: string;
  generationId: number;
}

export type GenerateEmailBodyTone =
  | (typeof GenerateEmailBodyTone)[keyof typeof GenerateEmailBodyTone]
  | null;

export const GenerateEmailBodyTone = {
  professional: "professional",
  conversational: "conversational",
  bold: "bold",
  empathetic: "empathetic",
  urgent: "urgent",
} as const;

export interface GenerateEmailBody {
  name: string;
  company: string;
  role: string;
  productType: string;
  problemHypothesis: string;
  recentHook: string;
  context: string;
  tone?: GenerateEmailBodyTone;
  playbookId?: number | null;
}

export interface GenerateScriptBody {
  objective: string;
  context: string;
  playbookId?: number | null;
}

export interface GenerateScriptBuilderBody {
  name?: string;
  company?: string;
  role?: string;
  productType?: string;
  scriptType?: string;
  context?: string;
  playbookId?: number | null;
  notes?: string[];
}

export interface SocialProofMeta {
  orgName: string;
  orgType: string;
  sizeTier: string;
  region: string;
  angle: string;
  intensity: string;
  matchReason: string;
  orgId: string;
  hasOHP: boolean;
}

export interface ScriptBuilderResult {
  output: string;
  generationId: number;
  socialProof?: SocialProofMeta | null;
}

export interface SuggestEditsBody {
  draftEmail: string;
  playbookId?: number | null;
}

export interface DashboardStats {
  totalPlaybooks: number;
  totalGenerations: number;
  totalEmailsAnalyzed: number;
  emailsGenerated: number;
  scriptsGenerated: number;
  editsRequested: number;
}

export interface ApiError {
  error: string;
}

export interface KnowledgeDoc {
  id: number;
  title: string;
  type: string;
  content: string;
  fileName?: string | null;
  fileType?: string | null;
  fileSize?: number | null;
  playbookId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateKnowledgeDocBody {
  title: string;
  type: string;
  content: string;
  playbookId?: number | null;
}

export interface MeetingSlot {
  id: string;
  address: string;
  scheduledStartTime: string;
  durationMinutes: number;
  companyId?: string | null;
  companyName?: string | null;
  lat?: number | null;
  lng?: number | null;
}

export type DirectionLegOrigin = {
  address: string;
  lat: number;
  lng: number;
};

export type DirectionLegDestination = {
  address: string;
  lat: number;
  lng: number;
};

export interface DirectionLeg {
  origin: DirectionLegOrigin;
  destination: DirectionLegDestination;
  distanceMeters: number;
  durationSeconds: number;
  polyline?: string | null;
}

export type ItineraryHome = {
  address: string;
  lat: number;
  lng: number;
} | null;

export interface Itinerary {
  date: string;
  home?: ItineraryHome;
  meetings: MeetingSlot[];
  totalDistanceMeters: number;
  totalDurationSeconds: number;
  directions: DirectionLeg[];
}

export interface ConfirmAddressBody {
  rawAddress: string;
  companyName?: string | null;
}

export interface AddressCandidate {
  address: string;
  lat: number;
  lng: number;
}

export interface ConfirmAddressResult {
  candidates: AddressCandidate[];
}

export type ListKnowledgeDocsParams = {
  playbookId?: number;
};

export type UploadKnowledgeDocBody = {
  file: Blob;
  title: string;
  type: string;
  playbookId?: number;
};

export type OptimizeTravelParams = {
  date?: string;
  homeBase?: string;
  seedAccountId?: string;
};