# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /playbooks:
    get:
      operationId: listPlaybooks
      tags: [playbooks]
      summary: List all playbooks
      responses:
        "200":
          description: List of playbooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Playbook"
    post:
      operationId: createPlaybook
      tags: [playbooks]
      summary: Create a new playbook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlaybookBody"
      responses:
        "201":
          description: Created playbook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playbook"
  /playbooks/{id}:
    get:
      operationId: getPlaybook
      tags: [playbooks]
      summary: Get a playbook with its patterns
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Playbook with patterns
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybookWithPatterns"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deletePlaybook
      tags: [playbooks]
      summary: Delete a playbook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /playbooks/{id}/analyze:
    post:
      operationId: analyzeEmails
      tags: [playbooks]
      summary: Analyze emails to extract patterns into a playbook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeEmailsBody"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /travel/optimize-travel:
    get:
      operationId: optimizeTravel
      tags: [travel]
      summary: Optimize travel itinerary for a given date
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
        - name: homeBase
          in: query
          required: false
          schema:
            type: string
        - name: seedAccountId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Optimized travel itinerary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Itinerary"
        "404":
          description: No meetings found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Geocoding error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /travel/confirm-address:
    post:
      operationId: confirmAddress
      tags: [travel]
      summary: Confirm and normalize a raw address string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmAddressBody"
      responses:
        "200":
          description: Address candidates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfirmAddressResult"
...
components:
  schemas:
...
    Playbook:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        qualityScore:
          type: integer
          nullable: true
        emailCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - emailCount
        - createdAt
        - updatedAt
    PlaybookWithPatterns:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        qualityScore:
          type: integer
          nullable: true
        emailCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        patterns:
          type: array
          items:
            $ref: "#/components/schemas/Pattern"
        principles:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - description
        - emailCount
        - createdAt
        - updatedAt
        - patterns
        - principles
    Pattern:
      type: object
      properties:
        id:
          type: integer
        playbookId:
          type: integer
        type:
          type: string
        text:
          type: string
        examples:
          type: array
          items:
            type: string
      required:
        - id
        - playbookId
        - type
        - text
        - examples
    CreatePlaybookBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        icpVerticals:
          type: array
          items:
            type: string
        icpPersonas:
          type: array
          items:
            type: string
        icpPainPoints:
          type: array
          items:
            type: string
        icpDifferentiators:
          type: array
          items:
            type: string
        icpProofPoints:
          type: array
          items:
            type: string
        icpCompanySize:
          type: string
          nullable: true
      required:
        - name
        - description
    AnalyzeEmailsBody:
      type: object
      properties:
        emails:
          type: array
          items:
            type: string
      required:
        - emails
    AnalysisResult:
      type: object
      properties:
        patterns:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              text:
                type: string
              examples:
                type: array
                items:
                  type: string
            required:
              - type
              - text
              - examples
        principles:
          type: array
          items:
            type: string
        qualityScore:
          type: integer
      required:
        - patterns
        - principles
        - qualityScore
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    MeetingSlot:
      type: object
      properties:
        id:
          type: string
        address:
          type: string
        scheduledStartTime:
          type: string
          format: date-time
        durationMinutes:
          type: number
        companyId:
          type: string
          nullable: true
        companyName:
          type: string
          nullable: true
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
      required:
        - id
        - address
        - scheduledStartTime
        - durationMinutes
    DirectionLeg:
      type: object
      properties:
        origin:
          type: object
          properties:
            address:
              type: string
            lat:
              type: number
            lng:
              type: number
          required:
            - address
            - lat
            - lng
        destination:
          type: object
          properties:
            address:
              type: string
            lat:
              type: number
            lng:
              type: number
          required:
            - address
            - lat
            - lng
        distanceMeters:
          type: number
        durationSeconds:
          type: number
        polyline:
          type: string
          nullable: true
      required:
        - origin
        - destination
        - distanceMeters
        - durationSeconds
    Itinerary:
      type: object
      properties:
        date:
          type: string
        home:
          type: object
          nullable: true
          properties:
            address:
              type: string
            lat:
              type: number
            lng:
              type: number
          required:
            - address
            - lat
            - lng
        meetings:
          type: array
          items:
            $ref: "#/components/schemas/MeetingSlot"
        totalDistanceMeters:
          type: number
        totalDurationSeconds:
          type: number
        directions:
          type: array
          items:
            $ref: "#/components/schemas/DirectionLeg"
      required:
        - date
        - meetings
        - totalDistanceMeters
        - totalDurationSeconds
        - directions
    ConfirmAddressBody:
      type: object
      properties:
        rawAddress:
          type: string
        companyName:
          type: string
          nullable: true
      required:
        - rawAddress
    AddressCandidate:
      type: object
      properties:
        address:
          type: string
        lat:
          type: number
        lng:
          type: number
      required:
        - address
        - lat
        - lng
    ConfirmAddressResult:
      type: object
      properties:
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/AddressCandidate"
      required:
        - candidates